# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Swagger API
basePath: /api
schemes:
  - http
  - https
paths:
  /bundles:
    post:
      description: Create a new bundle version from the zip file
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: bundleName
          in: formData
          required: true
          type: string
          description: The name of the bundle
        - name: bundleData
          in: formData
          type: file
          required: true
          description: The data for the bundle file
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/BundleCreated'
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to post this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
  /bundles/{bundleName}/revisions:
    parameters:
      - $ref: '#/parameters/bundleName'
      - $ref: '#/parameters/cursor'
    get:
      description: Get all revisions of he bundle.
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        201:
          description: Success
          schema:
            $ref: '#/definitions/BundleRevisions'
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to post this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
  /bundles/{bundleName}/revisions/{bundleRevision}:
    parameters:
      - $ref: '#/parameters/bundleName'
      - $ref: '#/parameters/bundleRevision'
    get:
      description: Retrieve the bundle.  Expects a bearer token in the header.
      produces:
        - application/octet-stream
        - application/zip
      responses:
        200:
          schema:
            type: file
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
    delete:
      description: Delete the bundle.  Expects a bearer token in the header
      responses:
        200:
          schema:
            type: file
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
  /bundles/{bundleName}/tags:
    parameters:
      - $ref: '#/parameters/bundleName'
      - $ref: '#/parameters/cursor'
    post:
      parameters:
        - name: _
          in: body
          required: true
          description: Input for tag creation
          schema:
            $ref: '#/definitions/TagCreate'
      description: Create a new tag with the bundle revision.
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/TagInfo'
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
    get:
      description: Get all tags for the bundle
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/Tags'
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
  /bundles/{bundleName}/tags/{tagName}:
    parameters:
      - $ref: '#/parameters/bundleName'
      - $ref: '#/parameters/tagName'
    get:
      description: Return the bundle download
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/TagInfo'
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
    delete:
      description: Return the bundle download
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/TagInfo'
          description: Success
        404:
          description: Bundle not found
        401:
          description: Not a valid JWT token
        403:
          description: You are not authorized to get this bundle
        default:
          description: Error
          schema:
            $ref:  "#/definitions/Errors"
definitions:
  Resource:
    type: object
    required:
      - self
    properties:
      self:
        type: string
        readOnly: true
  CollectionResponse:
    allOf:
    - $ref: '#/definitions/Resource'
    required:
      - cursor
    properties:
      cursor:
        type: string
        readOnly: true
  BundleCreated:
    allOf:
    - $ref: '#/definitions/Resource'
    properties:
      revision:
        type: string
        description: A revision that represents a unique version of this bundle
  BundleRevisions:
    allOf:
    - $ref: '#/definitions/CollectionResponse'
    properties:
      revisions:
        type: array
        items:
          $ref: '#/definitions/RevisionEntry'
        description: A list of all urls for this bundle
  RevisionEntry:
    allOf:
    - $ref: '#/definitions/Resource'
    properties:
      revision:
        type: string
        description: A list of all urls for this bundle
      date:
        type: string
        description: A list of all urls for this bundle
        format: date-time
  Tags:
    allOf:
    - $ref: '#/definitions/CollectionResponse'
    properties:
      tags:
        type: array
        items:
          $ref: '#/definitions/TagInfo'
  TagInfo:
    allOf:
    - $ref: '#/definitions/Resource'
    properties:
      tag:
        type: string
        description: The name of a tag
      revision:
        type: string
        description: The revision in the bundle to set in the tag
      bundleUrl:
        type: string
        description: The url of the bundle
  TagCreate:
    properties:
      tag:
        type: string
        description: The name of a tag
      revision:
        type: string
        description: The revision in the bundle to set in the tag
  Errors:
    properties:
       errors:
        type: array
        items:
          type: string
parameters:
  bundleName:
    name: bundleName
    in: path
    required: true
    description: The name
    type: string
  bundleRevision:
    name: bundleRevision
    in: path
    required: true
    description: The revision of the bundle
    type: string
  tagName:
    name: tagName
    in: path
    required: true
    description: The name of the tag
    type: string
  cursor:
    name: cursor
    in: query
    required: false
    description: The cursor used to
    type: string